backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
dj-database-url==3.0.1
Pillow==10.4.0
